+++ packages/plugins/type-distributor/src/headless/stores/ssd-import.store.svelte.ts
@@ -13,6 +13,7 @@ import {
 	parseTemplates,
 	parseDataTypeTemplates
 } from '@/headless/ssd-parsing'
+import { bayTypesStore } from './bay-types.store.svelte'
 
 class UseImportSSDStore {
 	fileInput = $state<HTMLInputElement>()
@@ -34,6 +35,8 @@ class UseImportSSDStore {
 	selectedBayType = $state<string | null>(null)
 
 	loadFromSSD(xmlDocument: XMLDocument, filename: string) {
+		bayTypesStore.clearCache()
+
 		this.bayTypes = parseBayTypes(xmlDocument)
 
 		const templates = parseTemplates(xmlDocument)