+++ packages/plugins/type-distributor/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.2] - 2026-02-18
 ## Added
 - Dialog notifying the User to open only .scd files initially
+### Changed
+- Restrict Switching Bays when working on assigning a BayType to a Bay.
+- Cache BayType resolution
 ### Fixed
 - Dropdown being cut into on scroll of right container